+++ lib/Request.php
@@ -126,6 +126,13 @@ class Request extends Base implements \ArrayAccess, \Countable, \IteratorAggrega
     protected $method;
 
     /**
+     * The extra keys course by &offsetGet
+     *
+     * @var array
+     */
+    protected $extraKeys = [];
+
+    /**
      * Constructor
      *
      * @param array $options
@@ -162,7 +169,7 @@ class Request extends Base implements \ArrayAccess, \Countable, \IteratorAggrega
      */
     public function __invoke($name, $default = '')
     {
-        return isset($this->data[$name]) ? (string)$this->data[$name] : $default;
+        return isset($this->data[$name]) ? (string) $this->data[$name] : $default;
     }
 
     /**
@@ -187,7 +194,7 @@ class Request extends Base implements \ArrayAccess, \Countable, \IteratorAggrega
      */
     public function getInt($name, $min = null, $max = null)
     {
-        $value = (int)$this($name);
+        $value = (int) $this($name);
 
         if (!is_null($min) && $value < $min) {
             return $min;
@@ -277,6 +284,9 @@ class Request extends Base implements \ArrayAccess, \Countable, \IteratorAggrega
      */
     public function &offsetGet($offset)
     {
+        if (!isset($this->data[$offset])) {
+            $this->extraKeys[$offset] = true;
+        }
         return $this->data[$offset];
     }
 
@@ -289,6 +299,7 @@ class Request extends Base implements \ArrayAccess, \Countable, \IteratorAggrega
      */
     public function offsetSet($offset, $value)
     {
+        unset($this->extraKeys[$offset]);
         return $this->data[$offset] = $value;
     }
 
@@ -299,6 +310,7 @@ class Request extends Base implements \ArrayAccess, \Countable, \IteratorAggrega
      */
     public function offsetUnset($offset)
     {
+        unset($this->extraKeys[$offset]);
         unset($this->data[$offset]);
     }
 
@@ -315,12 +327,23 @@ class Request extends Base implements \ArrayAccess, \Countable, \IteratorAggrega
     }
 
     /**
-     * Return the source array of the object
+     * Returns the request parameters
      *
      * @return array
      */
     public function toArray()
     {
+        return $this->getData();
+    }
+
+    /**
+     * Returns the request parameters
+     *
+     * @return array
+     */
+    public function getData()
+    {
+        $this->removeExtraKeys();
         return $this->data;
     }
 
@@ -331,7 +354,7 @@ class Request extends Base implements \ArrayAccess, \Countable, \IteratorAggrega
      */
     public function count()
     {
-        return count($this->data);
+        return count($this->getData());
     }
 
     /**
@@ -612,8 +635,8 @@ class Request extends Base implements \ArrayAccess, \Countable, \IteratorAggrega
             $header .= $name . ': ' . $value . "\r\n";
         }
         return $this->getServer('REQUEST_METHOD') . ' ' . $this->getUrl() . ' ' . $this->getServer('SERVER_PROTOCOL') . "\r\n"
-        . $header
-        . $this->getContent();
+            . $header
+            . $this->getContent();
     }
 
     /**
@@ -894,7 +917,7 @@ class Request extends Base implements \ArrayAccess, \Countable, \IteratorAggrega
      */
     public function getIterator()
     {
-        return new \ArrayIterator($this->data);
+        return new \ArrayIterator($this->getData());
     }
 
     /**
@@ -929,5 +952,18 @@ class Request extends Base implements \ArrayAccess, \Countable, \IteratorAggrega
     public function getReferer()
     {
         return $this->getServer('HTTP_REFERER');
+    }
+
+    /**
+     * Removes extra keys in data
+     */
+    protected function removeExtraKeys()
+    {
+        foreach ($this->extraKeys as $offset => $value) {
+            if ($this->data[$offset] === null) {
+                unset($this->data[$offset]);
+            }
+        }
+        $this->extraKeys = [];
     }
 }